3.- Aula

+++ src/main/java/org/iesalandalus/programacion/reservasaulas/mvc/modelo/dominio/Aula.java
@@ -36,7 +36,7 @@ public class Aula {
 	}
 
 	
-	public int getPuntos() {
+	public float getPuntos() {
 		return puestos * PUNTOSXPUESTO;
 	}
